Reduce .v21 and .v30 to minimal differences between the two IMs
- Implement .common.Base classes for those which differ in attributes
  between the two IMs: BaseDataflow, BaseObservation, BaseDataset,
  BaseMetadataStructureDefinition, BaseMetadataflow, BaseSelectionValue,
  BaseMemberValue, BaseDataKeySet, BaseMemberSelection
- Implement BaseConstraint.__contains__ for typing.
- Order and number code sections to match the respective standards.
- Move ConstrainableArtefact to §3.2; allows better ordering.
- Use "(Meta)Dataflow" (SDMX 3.0 class names) instead of "-Definition"
  (2.1) for brevity.
